Run without error if spring is not installed
We have spring in the development and test groups which means that it's
present in the Gemfile but not installed on production. We are also not
using the RAILS_ENV variable when logging in as developer and start a
console with a parameter:

    ./bin/rails console -e production

In this case, a LoadError was raised when trying to load spring. So
since checking the RAILS_ENV is not a reliable detection of spring being
installed, we added a rescue call.